Bump deps, export elections
