+++ package.json
@@ -29,10 +29,10 @@
     "test:watch": "jest --coverage --watch"
   },
   "devDependencies": {
-    "@babel/core": "^7.5.4",
-    "@babel/register": "^7.4.4",
-    "@babel/runtime": "^7.5.4",
-    "@polkadot/dev": "^0.30.0-beta.22",
+    "@babel/core": "^7.5.5",
+    "@babel/register": "^7.5.5",
+    "@babel/runtime": "^7.5.5",
+    "@polkadot/dev": "^0.30.0-beta.23",
     "@polkadot/ts": "^0.1.62",
     "gh-pages": "^2.0.1"
   }
