+++ packages/api-derive/package.json
@@ -27,13 +27,13 @@
   },
   "homepage": "https://github.com/polkadot-js/api/tree/master/packages/api-derive#readme",
   "dependencies": {
-    "@babel/runtime": "^7.5.4",
+    "@babel/runtime": "^7.5.5",
     "@polkadot/api": "^0.82.0-beta.71",
     "@polkadot/types": "^0.82.0-beta.71",
     "@types/memoizee": "^0.4.2",
     "memoizee": "^0.4.14"
   },
   "devDependencies": {
-    "@polkadot/keyring": "^0.94.0-beta.15"
+    "@polkadot/keyring": "^0.94.0-beta.16"
   }
 }
