+++ packages/api-derive/src/elections/voters.ts
@@ -1,3 +1,7 @@
+// Copyright 2017-2019 @polkadot/api-derive authors & contributors
+// This software may be modified and distributed under the terms
+// of the Apache-2.0 license. See the LICENSE file for details.
+
 import { Observable } from 'rxjs';
 import { switchMap, map } from 'rxjs/operators';
 import { ApiInterfaceRx } from '@polkadot/api/types';
