+++ packages/api-derive/src/util/index.ts
@@ -1,0 +1,5 @@
+// Copyright 2017-2019 @polkadot/api-derive authors & contributors
+// This software may be modified and distributed under the terms
+// of the Apache-2.0 license. See the LICENSE file for details.
+
+export * from './approvalFlagToBool';
