+++ packages/api-metadata/package.json
@@ -26,12 +26,12 @@
   },
   "homepage": "https://github.com/polkadot-js/api/tree/master/packages/type-metadata#readme",
   "dependencies": {
-    "@babel/runtime": "^7.5.4",
+    "@babel/runtime": "^7.5.5",
     "@polkadot/types": "^0.82.0-beta.71",
-    "@polkadot/util": "^0.94.0-beta.15",
-    "@polkadot/util-crypto": "^0.94.0-beta.15"
+    "@polkadot/util": "^0.94.0-beta.16",
+    "@polkadot/util-crypto": "^0.94.0-beta.16"
   },
   "devDependencies": {
-    "@polkadot/keyring": "^0.94.0-beta.15"
+    "@polkadot/keyring": "^0.94.0-beta.16"
   }
 }
