+++ packages/rpc-core/package.json
@@ -26,11 +26,11 @@
   },
   "homepage": "https://github.com/polkadot-js/api/tree/master/packages/rpc-core#readme",
   "dependencies": {
-    "@babel/runtime": "^7.5.4",
+    "@babel/runtime": "^7.5.5",
     "@polkadot/jsonrpc": "^0.82.0-beta.71",
     "@polkadot/rpc-provider": "^0.82.0-beta.71",
     "@polkadot/types": "^0.82.0-beta.71",
-    "@polkadot/util": "^0.94.0-beta.15",
+    "@polkadot/util": "^0.94.0-beta.16",
     "memoizee": "^0.4.14",
     "rxjs": "^6.5.2"
   }
