+++ packages/rpc-provider/package.json
@@ -26,17 +26,17 @@
   },
   "homepage": "https://github.com/polkadot-js/api/tree/master/packages/rpc-provider#readme",
   "dependencies": {
-    "@babel/runtime": "^7.5.4",
+    "@babel/runtime": "^7.5.5",
     "@polkadot/api-metadata": "^0.82.0-beta.71",
-    "@polkadot/util": "^0.94.0-beta.15",
-    "@polkadot/util-crypto": "^0.94.0-beta.15",
+    "@polkadot/util": "^0.94.0-beta.16",
+    "@polkadot/util-crypto": "^0.94.0-beta.16",
     "@types/nock": "^10.0.3",
     "eventemitter3": "^4.0.0",
     "isomorphic-fetch": "^2.2.1",
     "websocket": "^1.0.29"
   },
   "devDependencies": {
-    "@polkadot/keyring": "^0.94.0-beta.15",
+    "@polkadot/keyring": "^0.94.0-beta.16",
     "mock-socket": "^8.0.5",
     "nock": "^10.0.4"
   }
