+++ packages/types/package.json
@@ -26,11 +26,11 @@
   },
   "homepage": "https://github.com/polkadot-js/api/tree/master/packages/types#readme",
   "dependencies": {
-    "@babel/runtime": "^7.5.4",
-    "@polkadot/util": "^0.94.0-beta.15",
-    "@polkadot/util-crypto": "^0.94.0-beta.15"
+    "@babel/runtime": "^7.5.5",
+    "@polkadot/util": "^0.94.0-beta.16",
+    "@polkadot/util-crypto": "^0.94.0-beta.16"
   },
   "devDependencies": {
-    "@polkadot/keyring": "^0.94.0-beta.15"
+    "@polkadot/keyring": "^0.94.0-beta.16"
   }
 }
